requirements.txt
-i https://pypi.org/simple
aiokafka==0.10.0
amqp==5.2.0 ; python_version >= '3.6'
annotated-types==0.6.0 ; python_version >= '3.8'
anyio==4.3.0 ; python_version >= '3.8'
asgiref==3.8.1 ; python_version >= '3.8'
async-timeout==4.0.3 ; python_version >= '3.7'
billiard==4.2.0 ; python_version >= '3.7'
celery==5.4.0
certifi==2024.2.2 ; python_version >= '3.6'
cffi==1.16.0 ; platform_python_implementation != 'PyPy'
click==8.1.7 ; python_version >= '3.7'
click-didyoumean==0.3.1 ; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0 ; python_version >= '3.6'
cryptography==42.0.5
dj-database-url==2.1.0
django==5.0.4
django-allauth==0.62.1
django-async-redis==0.2.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-ninja==1.1.0
django-redis==5.4.0
fast-depends==2.4.2 ; python_version >= '3.8'
faststream[kafka]==0.5.3
gunicorn==22.0.0
h11==0.14.0 ; python_version >= '3.7'
idna==3.7 ; python_version >= '3.5'
kombu==5.3.7 ; python_version >= '3.8'
markdown-it-py==3.0.0 ; python_version >= '3.8'
mdurl==0.1.2 ; python_version >= '3.7'
packaging==24.0 ; python_version >= '3.7'
prompt-toolkit==3.0.43 ; python_full_version >= '3.7.0'
psycopg2-binary==2.9.9
pycparser==2.22 ; python_version >= '3.8'
pydantic==2.7.1 ; python_version >= '3.8'
pydantic-core==2.18.2 ; python_version >= '3.8'
pygments==2.17.2 ; python_version >= '3.7'
pyjwt[crypto]==2.8.0
python-dateutil==2.9.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
redis==5.0.4 ; python_version >= '3.7'
rich==13.7.1 ; python_full_version >= '3.7.0'
sentry-sdk==2.0.1
shellingham==1.5.4 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1 ; python_version >= '3.7'
sqlparse==0.5.0 ; python_version >= '3.8'
typer==0.12.3 ; python_version >= '3.7'
typing-extensions==4.11.0 ; python_version >= '3.8'
tzdata==2024.1 ; python_version >= '2'
urllib3==2.2.1 ; python_version >= '3.8'
uvicorn==0.29.0
vine==5.1.0 ; python_version >= '3.6'
wcwidth==0.2.13
whitenoise==6.6.0
